app/requirements.txt
langchain==0.0.316
openai==0.28.1
streamlit==1.28.0
streamlit-feedback==0.1.2
requests==2.31.0
requests-file==1.5.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
tiktoken==0.5.1